Integrate Advanced Analytics into ProgressPage as tabs instead of modal, fix error handling and improve UX
